Time: 4 ms (100.00%), Space: 15.2 MB (100.00%) - LeetHub

+++ 1220-count-vowels-permutation/1220-count-vowels-permutation.c
@@ -5,7 +5,7 @@
 #define u 4
 #define MOD(a) ((a)%1000000007ULL)
 
-unsigned long long countVowelPermutation(int n) {
+int countVowelPermutation(int n) {
     int idx;
     unsigned long long *dp[5];
